fix(auth): correct oauth2 endpoints and improve flow
